requirements.txt
Django==3.2
drf-yasg==1.21.5
djangorestframework==3.14.0
pandas==2.0.1
numpy==1.24.3